fix: chat list overflow-audo should not go beyond parent
