+++ src/index.css
@@ -14,6 +14,7 @@ https://benborgers.com/posts/tailwind-h-screen
 
 ::-webkit-scrollbar {
     width: 3px;
+    height: 4px;
 }
 
 ::-webkit-scrollbar-thumb {
