+++ src/panel/panel-example.tsx
