+++ src/panel/setting/ability/ability-setting.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import {ClientAbility} from "../../../../../state/data-structure/client-ability/client-ability.tsx";
+import {ClientAbility} from "../../../state/data-structure/client-ability/client-ability.tsx";
 import ChatGpt from "./chat-gpt.tsx";
 
 type Props = {
