+++ src/panel/setting/widget/rainbow.ts
