+++ src/panel/setting/widget/switch.tsx
